requirements.txt
customtkinter==5.1.2
music_tag==0.4.3
openai==0.27.4
Pillow==9.5.0
